% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_some_numbers.R
\name{addition}
\alias{addition}
\title{Add some stuff together}
\usage{
addition(x, y)
}
\arguments{
\item{x}{a number [=] real}

\item{y}{another number [=] real}
}
\value{
the sum of x and y
}
\description{
Add some stuff together
}
