% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{onLoad function for sean's r package}
\usage{
.onLoad(libname, pkgname)
}
\description{
This function loads some default values into the environment, 
The file is called zzz.R because it will be the last file loaded upon a library load. 
This is traditional to R package design.
}
