% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_functions_rand_numbers_and_plot.R
\name{make_random_normal_variables}
\alias{make_random_normal_variables}
\title{Make some random normal variables}
\usage{
make_random_normal_variables(size = 100)
}
\arguments{
\item{size}{is the size of population [=] integer}
}
\value{
set of random normal variables
}
\description{
Make some random normal variables
}
