% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_functions_rand_numbers_and_plot.R
\name{plot_some_random_stuff}
\alias{plot_some_random_stuff}
\title{Plot a histogram}
\usage{
plot_some_random_stuff(vector_of_numbers)
}
\arguments{
\item{vector_of_numbers}{to be plotted [=] a vector}
}
\value{
graphic of some random numbers in a histogram form, to the console
}
\description{
Plot a histogram
}
