% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints_something.R
\name{print_some_stuff}
\alias{print_some_stuff}
\title{Print a phrase}
\usage{
print_some_stuff()
}
\arguments{
\item{NULL}{}
}
\value{
prints phrase to the console
}
\description{
prints something everyone will know
}
